+++ src/main/kotlin/net/trivernis/chunkmaster/lib/generation/ChunkCoordinates.kt
@@ -1,0 +1,4 @@
+package net.trivernis.chunkmaster.lib.generation
+
+data class ChunkCoordinates(val x: Int, val z: Int) {
+}
