+++ src/main/kotlin/net/trivernis/chunkmaster/lib/generation/GenerationTaskPaper.kt
@@ -1,25 +1,30 @@
-package net.trivernis.chunkmaster.lib
+package net.trivernis.chunkmaster.lib.generation
 
 import net.trivernis.chunkmaster.Chunkmaster
 import org.bukkit.Chunk
 import org.bukkit.World
-import org.bukkit.scheduler.BukkitRunnable
-import org.bukkit.scheduler.BukkitTask
 import java.util.concurrent.CompletableFuture
+import io.papermc.lib.PaperLib
 
-class GenerationTaskSpigot(
+class GenerationTaskPaper(
     private val plugin: Chunkmaster, override val world: World,
     centerChunk: Chunk, private val startChunk: Chunk,
     override val stopAfter: Int = -1
 ) : GenerationTask(plugin, centerChunk, startChunk) {
 
+    private val pendingChunks = HashSet<CompletableFuture<Chunk>>()
 
     override var count = 0
         private set
     override var lastChunk: Chunk = startChunk
+        get() {
+            return world.getChunkAt(lastChunkCoords.x, lastChunkCoords.z)
+        }
         private set
     override var endReached: Boolean = false
         private set
+
+    private var lastChunkCoords = ChunkCoordinates(startChunk.x, startChunk.z)
 
     /**
      * Runs the generation task. Every Iteration the next chunk will be generated if
@@ -34,30 +39,29 @@ class GenerationTaskSpigot(
                         chunk.unload(true)
                     }
                 }
-            } else {
+            } else if (pendingChunks.size < 10) {   // if more than 10 chunks are pending, wait.
                 if (borderReached()) {
                     endReached = true
                     return
                 }
 
-                var chunk = nextChunk
-
+                var chunk = nextChunkCoordinates
                 for (i in 1 until chunkSkips) {
-                    if (world.isChunkGenerated(chunk.x, chunk.z)) {
-                        chunk = nextChunk
+                    if (PaperLib.isChunkGenerated(world, chunk.x, chunk.z)) {
+                        chunk = nextChunkCoordinates
                     } else {
                         break
                     }
                 }
 
-                if (!world.isChunkGenerated(chunk.x, chunk.z)) {
-                    chunk.load(true)
-                    loadedChunks.add(chunk)
+                if (!PaperLib.isChunkGenerated(world, chunk.x, chunk.z)) {
+                    pendingChunks.add(PaperLib.getChunkAtAsync(world, chunk.x, chunk.z, true))
                 }
-                lastChunk = chunk
+                lastChunkCoords = chunk
                 count = spiral.count // set the count to the more accurate spiral count
             }
         }
+        checkChunksLoaded()
     }
 
     /**
@@ -65,10 +69,34 @@ class GenerationTaskSpigot(
      * This unloads all chunks that were generated but not unloaded yet.
      */
     override fun cancel() {
+        for (pendingChunk in pendingChunks) {
+            if (pendingChunk.isDone) {
+                loadedChunks.add(pendingChunk.get())
+            } else {
+                pendingChunk.cancel(true)
+            }
+        }
+        pendingChunks.clear()
         for (chunk in loadedChunks) {
             if (chunk.isLoaded) {
                 chunk.unload(true)
             }
         }
+    }
+
+    /**
+     * Checks if some chunks have been loaded and adds them to the loaded chunk set.
+     */
+    private fun checkChunksLoaded() {
+        val completedEntrys = HashSet<CompletableFuture<Chunk>>()
+        for (pendingChunk in pendingChunks) {
+            if (pendingChunk.isDone) {
+                completedEntrys.add(pendingChunk)
+                loadedChunks.add(pendingChunk.get())
+            } else if (pendingChunk.isCompletedExceptionally || pendingChunk.isCancelled) {
+                completedEntrys.add(pendingChunk)
+            }
+        }
+        pendingChunks.removeAll(completedEntrys)
     }
 }
