+++ src/main/kotlin/net/trivernis/chunkmaster/lib/generation/GenerationTaskSpigot.kt
@@ -1,18 +1,15 @@
-package net.trivernis.chunkmaster.lib
+package net.trivernis.chunkmaster.lib.generation
 
 import net.trivernis.chunkmaster.Chunkmaster
 import org.bukkit.Chunk
 import org.bukkit.World
-import java.util.concurrent.CompletableFuture
-import io.papermc.lib.PaperLib
 
-class GenerationTaskPaper(
+class GenerationTaskSpigot(
     private val plugin: Chunkmaster, override val world: World,
     centerChunk: Chunk, private val startChunk: Chunk,
     override val stopAfter: Int = -1
 ) : GenerationTask(plugin, centerChunk, startChunk) {
 
-    private val pendingChunks = HashSet<CompletableFuture<Chunk>>()
 
     override var count = 0
         private set
@@ -34,29 +31,30 @@ class GenerationTaskPaper(
                         chunk.unload(true)
                     }
                 }
-            } else if (pendingChunks.size < 10) {   // if more than 10 chunks are pending, wait.
+            } else {
                 if (borderReached()) {
                     endReached = true
                     return
                 }
 
                 var chunk = nextChunk
+
                 for (i in 1 until chunkSkips) {
-                    if (PaperLib.isChunkGenerated(world, chunk.x, chunk.z)) {
+                    if (world.isChunkGenerated(chunk.x, chunk.z)) {
                         chunk = nextChunk
                     } else {
                         break
                     }
                 }
 
-                if (!PaperLib.isChunkGenerated(world, chunk.x, chunk.z)) {
-                    pendingChunks.add(PaperLib.getChunkAtAsync(world, chunk.x, chunk.z, true))
+                if (!world.isChunkGenerated(chunk.x, chunk.z)) {
+                    chunk.load(true)
+                    loadedChunks.add(chunk)
                 }
                 lastChunk = chunk
                 count = spiral.count // set the count to the more accurate spiral count
             }
         }
-        checkChunksLoaded()
     }
 
     /**
@@ -64,34 +62,10 @@ class GenerationTaskPaper(
      * This unloads all chunks that were generated but not unloaded yet.
      */
     override fun cancel() {
-        for (pendingChunk in pendingChunks) {
-            if (pendingChunk.isDone) {
-                loadedChunks.add(pendingChunk.get())
-            } else {
-                pendingChunk.cancel(true)
-            }
-        }
-        pendingChunks.clear()
         for (chunk in loadedChunks) {
             if (chunk.isLoaded) {
                 chunk.unload(true)
             }
         }
-    }
-
-    /**
-     * Checks if some chunks have been loaded and adds them to the loaded chunk set.
-     */
-    private fun checkChunksLoaded() {
-        val completedEntrys = HashSet<CompletableFuture<Chunk>>()
-        for (pendingChunk in pendingChunks) {
-            if (pendingChunk.isDone) {
-                completedEntrys.add(pendingChunk)
-                loadedChunks.add(pendingChunk.get())
-            } else if (pendingChunk.isCompletedExceptionally || pendingChunk.isCancelled) {
-                completedEntrys.add(pendingChunk)
-            }
-        }
-        pendingChunks.removeAll(completedEntrys)
     }
 }
